+++ brownford/src/main/java/com/brownford/repository/SectionCourseRepository.java
@@ -1,7 +1,0 @@
-package com.brownford.repository;
-
-import com.brownford.model.SectionCourse;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SectionCourseRepository extends JpaRepository<SectionCourse, Long> {
-}