+++ brownford/src/main/java/com/brownford/service/ScheduleService.java
@@ -1,7 +1,8 @@
 package com.brownford.service;
 
+import com.brownford.model.CurriculumCourse;
 import com.brownford.model.Schedule;
-import com.brownford.model.SectionCourse;
+import com.brownford.model.Section;
 import com.brownford.repository.ScheduleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,10 +15,6 @@ public class ScheduleService {
     @Autowired
     private ScheduleRepository scheduleRepository;
 
-    public List<Schedule> getSchedulesBySectionCourse(SectionCourse sectionCourse) {
-        return scheduleRepository.findBySectionCourse(sectionCourse);
-    }
-
     public Optional<Schedule> getScheduleById(Long id) {
         return scheduleRepository.findById(id);
     }
@@ -28,5 +25,26 @@ public class ScheduleService {
 
     public void deleteSchedule(Long id) {
         scheduleRepository.deleteById(id);
+    }
+
+    public List<Schedule> findByCurriculumCourseId(Long curriculumCourseId) {
+        CurriculumCourse cc = new CurriculumCourse();
+        cc.setId(curriculumCourseId);
+        return scheduleRepository.findByCurriculumCourse(cc);
+    }
+    public List<Schedule> findBySectionId(Long sectionId) {
+        Section section = new Section();
+        section.setId(sectionId);
+        return scheduleRepository.findBySection(section);
+    }
+    public List<Schedule> findByCurriculumCourseIdAndSectionId(Long curriculumCourseId, Long sectionId) {
+        CurriculumCourse cc = new CurriculumCourse();
+        cc.setId(curriculumCourseId);
+        Section section = new Section();
+        section.setId(sectionId);
+        return scheduleRepository.findByCurriculumCourseAndSection(cc, section);
+    }
+    public List<Schedule> getAllSchedules() {
+        return scheduleRepository.findAll();
     }
 }